Move `buy_complete_set` tests to new tests
